backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10